Move actions under the lib namespace
