+++ lib/cast_vote.js
@@ -1,12 +1,12 @@
 const Assert = require('assert-plus')
 const Vote = require('../entities/sys/vote')
-const Shapes = require('../lib/shapes')
-const { fetchProp } = require('../lib/utils')
-const { ValidationError, NotFoundError } = require('../lib/errors')
+const Shapes = require('./shapes')
+const { fetchProp } = require('./utils')
+const { ValidationError, NotFoundError } = require('./errors')
 const CastVoteService = require('../services/cast_vote')
 const GetVoteStats = require('../services/get_vote_stats')
 const SavePollRating = require('../services/save_poll_rating')
-const Reply = require('../lib/reply')
+const Reply = require('./reply')
 
 module.exports = function (opts = {}) {
   return async function (msg, reply) {
