+++ lib/get_poll.js
@@ -1,8 +1,8 @@
 const Assert = require('assert-plus')
-const Shapes = require('../lib/shapes')
-const { fetchProp } = require('../lib/utils')
-const { ValidationError } = require('../lib/errors')
-const Reply = require('../lib/reply')
+const Shapes = require('./shapes')
+const { fetchProp } = require('./utils')
+const { ValidationError } = require('./errors')
+const Reply = require('./reply')
 
 module.exports = function (opts = {}) {
   return async function (msg, reply) {
