+++ lib/open_poll.js
@@ -1,10 +1,10 @@
 const Assert = require('assert-plus')
 const Joi = require('joi')
-const Shapes = require('../lib/shapes')
-const { fetchProp } = require('../lib/utils')
-const { ValidationError } = require('../lib/errors')
+const Shapes = require('./shapes')
+const { fetchProp } = require('./utils')
+const { ValidationError } = require('./errors')
 const OpenPollService = require('../services/open_poll')
-const Reply = require('../lib/reply')
+const Reply = require('./reply')
 
 module.exports = function (opts = {}) {
   return async function (msg, reply) {
